# improve `givensAlgorithm(::AbstractFloat)`